Rescope data and other improvements

+++ packages/cx/src/data/ZoomIntoPropertyView.js
@@ -1,40 +1,31 @@
-import {View} from './View';
-import {Binding} from './Binding';
+import { View } from "./View";
+import { Binding } from "./Binding";
+import { AugmentedViewBase } from "./AugmentedViewBase";
+import { NestedDataView } from "./NestedDataView";
 
-export class ZoomIntoPropertyView extends View {
+export class ZoomIntoPropertyView extends NestedDataView {
+   getBaseData(parentStoreData) {
+      let x = this.binding.value(parentStoreData);
+      if (x != null && typeof x != "object") throw new Error("Zoomed value must be an object.");
+      return {
+         ...x,
+      };
+   }
 
-   getData() {
-      if (this.cache.version != this.meta.version) {
-         var data = this.store.getData();
-         var x = this.binding.value(data);
-         if (x != null && typeof x != 'object')
-            throw new Error('Zoomed value must be an object.');
-         this.cache.result = {
-            ...x,
-            [this.rootName]: !this.store.sealed ? {...data} : data
-         };
-         this.cache.version = this.meta.version;
-      }
-      return this.cache.result;
+   embedAugmentData(result, parentStoreData) {
+      result[this.rootName] = parentStoreData;
+      super.embedAugmentData(result, parentStoreData);
    }
 
    setItem(path, value) {
-      if (path.indexOf(this.rootName + '.') == 0)
-         this.store.setItem(path.substring(this.rootName.length + 1), value);
-      else
-         this.store.setItem(this.binding.path + '.' + path, value);
+      if (path.indexOf(this.rootName + ".") == 0) this.store.setItem(path.substring(this.rootName.length + 1), value);
+      else super.setItem(this.binding.path + "." + path, value);
    }
 
    deleteItem(path) {
-      if (path instanceof Binding)
-         path = path.path;
-
-      if (path.indexOf(this.rootName + '.') == 0)
-         this.store.deleteItem(path.substring(this.rootName.length + 1));
-      else
-         this.store.deleteItem(this.binding.path + '.' + path);
+      if (path.indexOf(this.rootName + ".") == 0) this.store.deleteItem(path.substring(this.rootName.length + 1));
+      else super.deleteItem(this.binding.path + "." + path);
    }
 }
 
-ZoomIntoPropertyView.prototype.rootName = '$root';
-
+ZoomIntoPropertyView.prototype.rootName = "$root";
